refactor and organize components
