+++ src/components/Footer.jsx
@@ -1,10 +1,9 @@
 import React from 'react';
-import HornedBeast from './HornedBeast'
 
 function Footer() {
     return (
       <footer>
-        <p>&copy; 2024 Evan Cheng</p>
+        <p>&copy;2024 Evan Cheng</p>
       </footer>
     );
   }
